create feature authentication and authorization

+++ app/Http/Controllers/Api/PostController.php
@@ -44,6 +44,7 @@ class PostController extends Controller
             $post = new Post();
             $post->title = $request->title;
             $post->description = $request->description;
+            $post->user_id = auth()->user()->id;
             $post->save();
 
             return response()->json([
@@ -61,7 +62,15 @@ class PostController extends Controller
         try {
             $post->title = $request->title;
             $post->description = $request->description;
-            $post->save();
+
+            if ($post->user_id === auth()->user()->id) {
+                $post->save();
+            } else {
+                return response()->json([
+                    'status_code' => 422,
+                    'status_message' => 'Anda tidak memiliki akses untuk mengubah',
+                ]);
+            }
 
             return response()->json([
                 'status_code' => 200,
@@ -76,7 +85,15 @@ class PostController extends Controller
     public function destroy(Post $post)
     {
         try {
-            $post->delete();
+            if ($post->user_id === auth()->user()->id) {
+                $post->delete();
+            } else {
+                return response()->json([
+                    'status_code' => 422,
+                    'status_message' => 'Anda tidak memiliki akses untuk mengubah',
+                ]);
+            }
+
             return response()->json([
                 'status_code' => 200,
                 'status_message' => 'Post berhasil dihapus',
